Fix the problem when using g++ 4.6 to build MCLinker

+++ include/mcld/MC/AttributeSet.h
@@ -34,7 +34,7 @@ public:
   typedef AttrSet::const_iterator const_iterator;
 
 public:
-  AttributeSet(size_t pNum, const Attribute& pPredefined);
+  AttributeSet(unsigned int pNum, const Attribute& pPredefined);
 
   ~AttributeSet();
